feat(svg): implement enhanced SVG sanitization with CSS tree integration
