+++ packages/enketo-core/package.json
@@ -52,6 +52,7 @@
     },
     "dependencies": {
         "bootstrap-datepicker": "^1.10.1",
+        "css-tree": "^3.2.1",
         "dompurify": "^3.4.5",
         "drag-drop-touch": "^1.3.1",
         "fix-webm-duration": "^1.0.6",
