+++ packages/enketo-core/src/js/utils.js
@@ -4,7 +4,7 @@
  * @module utils
  */
 
-import DOMPurify from 'dompurify';
+import { sanitizeSvg } from './sanitization-utils';
 
 let cookies;
 
@@ -289,225 +289,6 @@ function encodeHtmlEntities(text) {
         .replace(/</g, '&lt;')
         .replace(/>/g, '&gt;')
         .replace(/"/g, '&quot;');
-}
-
-const ALLOWED_SVG_STYLE_PROPERTIES = new Set([
-    'fill',
-    'fill-opacity',
-    'fill-rule',
-    'stroke',
-    'stroke-opacity',
-    'stroke-width',
-    'stroke-linecap',
-    'stroke-linejoin',
-    'stroke-miterlimit',
-    'stroke-dasharray',
-    'stroke-dashoffset',
-    'opacity',
-    'color',
-    'font-family',
-    'font-size',
-    'font-style',
-    'font-weight',
-    'text-anchor',
-    'letter-spacing',
-    'word-spacing',
-    'shape-rendering',
-    'text-rendering',
-    'image-rendering',
-    'transform',
-    'transform-origin',
-]);
-
-const BLOCKED_STYLE_VALUE =
-    /\b(url|var|expression)\s*\(|@import|javascript:|data:|vbscript:/i;
-
-const SAFE_TRANSFORM_VALUE =
-    /^(matrix|translate|scale|rotate|skewX|skewY)\([0-9eE+\-.,\s%deg]+\)(\s+(matrix|translate|scale|rotate|skewX|skewY)\([0-9eE+\-.,\s%deg]+\))*$/;
-
-function sanitizeStyleProp(propName, value) {
-    const normalizedPropName = propName.trim().toLowerCase();
-    let normalizedValue = value.trim();
-
-    if (!ALLOWED_SVG_STYLE_PROPERTIES.has(normalizedPropName)) {
-        return null;
-    }
-
-    if (!normalizedValue || BLOCKED_STYLE_VALUE.test(normalizedValue)) {
-        return null;
-    }
-
-    if (
-        normalizedPropName === 'transform' &&
-        !SAFE_TRANSFORM_VALUE.test(normalizedValue)
-    ) {
-        return null;
-    }
-
-    return {
-        propName: normalizedPropName,
-        value: normalizedValue,
-    };
-}
-
-function sanitizeStyleDeclarations(styles) {
-    const output = {};
-
-    Array.from(styles).forEach((propName) => {
-        const value = styles.getPropertyValue(propName).trim();
-        const sanitizedProp = sanitizeStyleProp(propName, value);
-
-        if (sanitizedProp) {
-            output[sanitizedProp.propName] = sanitizedProp.value;
-        }
-    });
-
-    return Object.entries(output)
-        .map(([propName, value]) => `${propName}:${value};`)
-        .join('');
-}
-
-function sanitizeInlineStyleText(styleText) {
-    if (!styleText) {
-        return '';
-    }
-
-    const output = {};
-
-    styleText.split(';').forEach((declaration) => {
-        const [propName, ...valueParts] = declaration.split(':');
-        const value = valueParts.join(':').trim();
-
-        if (!propName || !value) {
-            return;
-        }
-
-        const sanitizedProp = sanitizeStyleProp(propName, value);
-
-        if (sanitizedProp) {
-            output[sanitizedProp.propName] = sanitizedProp.value;
-        }
-    });
-
-    return Object.entries(output)
-        .map(([propName, value]) => `${propName}:${value};`)
-        .join('');
-}
-
-// Parses and separates the text content of a <style> element
-// into individual rules and declarations to sanitize them.
-function sanitizeStyleElementText(node) {
-    if (!node || !node.ownerDocument) {
-        return '';
-    }
-
-    const parsedCssDocument = document.implementation.createHTMLDocument('');
-    const parsedCssStyleElement = parsedCssDocument.createElement('style');
-    parsedCssStyleElement.textContent = node.textContent;
-    parsedCssDocument.head.appendChild(parsedCssStyleElement);
-
-    const output = [];
-
-    try {
-        const { cssRules } = parsedCssStyleElement.sheet;
-
-        Array.from(cssRules).forEach((rule) => {
-            // Keep selector rules only; drop @import, @font-face, @media and others.
-            if (rule.type === CSSRule.STYLE_RULE && rule.selectorText) {
-                const declarations = sanitizeStyleDeclarations(rule.style);
-
-                if (declarations) {
-                    output.push(`${rule.selectorText}{${declarations}}`);
-                }
-            }
-        });
-    } catch {
-        return '';
-    }
-
-    return output.join('\n');
-}
-
-// Tags that are forbidden when DOMPurify sanitizes SVG content.
-const DOMPURIFY_SVG_CONFIG = {
-    USE_PROFILES: { svg: true, svgFilters: true },
-    RETURN_DOM_FRAGMENT: true,
-    WHOLE_DOCUMENT: false,
-    FORBID_TAGS: [
-        'foreignObject',
-        'iframe',
-        'object',
-        'embed',
-        'link',
-        'meta',
-        'script',
-    ],
-};
-
-/**
- * Sanitizes an SVG element using DOMPurify, removing potentially dangerous
- * elements and attributes (scripts, event handlers, javascript: URLs, etc.).
- * @param {SVGElement} svgElement - The SVG element to sanitize
- * @return {SVGElement|null|undefined} The sanitized SVG element
- */
-function sanitizeSvg(svgElement) {
-    if (!svgElement) {
-        return svgElement;
-    }
-
-    // Strip <foreignObject> subtrees at the XML level before any HTML parsing
-    // occurs. The onerror payload is dispatched during DOMPurify's internal
-    // HTML parse step.
-    // Doing this pre-strip here ensures dangerous HTML descendants from
-    // foreignObject never reach that parse step.
-    // This should not happen since we're using an inert document, but we are keeping
-    // this as a safeguard to avoid any potential code execution from foreingObject descendants.
-    Array.from(
-        svgElement.getElementsByTagNameNS(
-            'http://www.w3.org/2000/svg',
-            'foreignObject'
-        )
-    ).forEach((el) => el.remove());
-
-    // Use an inert document to prevent any potential code execution on live document during the sanitization process.
-    const inertDoc = document.implementation.createHTMLDocument('');
-    const container = inertDoc.createElement('div');
-    container.appendChild(inertDoc.importNode(svgElement, true));
-
-    // Hooks to deal with sanitization of style elements and attributes, based on the approach described in
-    DOMPurify.addHook('uponSanitizeElement', (node, data) => {
-        if (data.tagName === 'style') {
-            node.textContent = sanitizeStyleElementText(node);
-        }
-    });
-
-    DOMPurify.addHook('afterSanitizeAttributes', (node) => {
-        if (node.hasAttribute && node.hasAttribute('style')) {
-            const sanitizedStyle = sanitizeInlineStyleText(
-                node.getAttribute('style')
-            );
-
-            if (sanitizedStyle) {
-                node.setAttribute('style', sanitizedStyle);
-            } else {
-                node.removeAttribute('style');
-            }
-        }
-    });
-
-    let fragment;
-
-    try {
-        fragment = DOMPurify.sanitize(
-            container.innerHTML,
-            DOMPURIFY_SVG_CONFIG
-        );
-    } finally {
-        DOMPurify.removeHook('uponSanitizeElement');
-        DOMPurify.removeHook('afterSanitizeAttributes');
-    }
-
-    return fragment.querySelector('svg');
 }
 
 export {
